Re-organize code to support more banks
