+++ bin/bank_scrap
@@ -1,5 +1,5 @@
 #!/usr/bin/env ruby
 $LOAD_PATH.unshift(File.expand_path(File.join(File.dirname(__FILE__), '..', 'lib')))
 
-require 'bankscrap'
+require 'bank_scrap'
 BankScrap::Cli.start(ARGV)
