+++ lib/bank_scrap.rb
@@ -1,0 +1,9 @@
+require 'bank_scrap/version'
+require 'bank_scrap/cli'
+require 'bank_scrap/bank'
+
+# TODO: load banks dynamically
+require 'bank_scrap/banks/bankinter'
+
+module BankScrap
+end
