+++ lib/bank_scrap/version.rb
