--- lib/bankscrap.rb
@@ -1,7 +1,0 @@
-require "bankscrap/version"
-require "bankinter/bankinter"
-require "bankscrap/cli"
-
-module BankScrap
-  # Your code goes here...
-end
